deploy/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
jinja2==3.1.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.3.0
azure-ai-textanalytics==5.3.0
email-validator==2.1.0